+++ Products/ATContentTypes/tests/test_installation.py
@@ -4,12 +4,8 @@ from Products.ATContentTypes.config import TOOLNAME
 from Products.ATContentTypes.tests import atcttestcase
 from Products.ATContentTypes.tool.atct import ATCTTool
 from Products.CMFCore.utils import getToolByName
-from Products.CMFPlone.utils import getFSVersionTuple
-
 import unittest
 
-
-PLONE5 = getFSVersionTuple()[0] >= 5
 tests = []
 
 
@@ -42,12 +38,16 @@ class TestInstallation(atcttestcase.ATCTSiteTestCase):
         # Test, if the Product is available in Quickinstaller.  For Plone < 5
         # it shouldn't (since it's a core dependency), for Plone >= 5 it should
         # (where plone.app.contenttypes is installed by default instead).
-        qi = getattr(self.portal, 'portal_quickinstaller')
-        products = [prod['id'] for prod in qi.listInstalledProducts()]
-        if PLONE5:
+        # Only Plone 5 is supported here.
+        try:
+            from Products.CMFPlone.utils import get_installer
+        except ImportError:
+            qi = getattr(self.portal, 'portal_quickinstaller')
+            products = [prod['id'] for prod in qi.listInstalledProducts()]
             self.assertTrue('ATContentTypes' in products)
         else:
-            self.assertTrue('ATContentTypes' not in products)
+            qi = get_installer(self.portal)
+            self.assertTrue(qi.is_product_installed('Products.ATContentTypes'))
 
     def test_release_settings_SAVE_TO_FAIL_FOR_DEVELOPMENT(self):
         self.assertEqual(SWALLOW_IMAGE_RESIZE_EXCEPTIONS, True)
@@ -72,6 +72,61 @@ class TestInstallation(atcttestcase.ATCTSiteTestCase):
         Products.ATContentTypes.atct  # pyflakes
 
 tests.append(TestInstallation)
+
+
+class TestUninstallation(atcttestcase.ATCTSiteTestCase):
+
+    def afterSetUp(self):
+        self.cat = getattr(self.portal.aq_explicit, 'portal_catalog')
+        try:
+            from Products.CMFPlone.utils import get_installer
+        except ImportError:
+            qi = getattr(self.portal, 'portal_quickinstaller')
+            qi.uninstallProducts(['ATContentTypes'])
+        else:
+            qi = get_installer(self.portal)
+            qi.uninstall_product('Products.ATContentTypes')
+
+    def test_tool_uninstalled(self):
+        t = getToolByName(self.portal, TOOLNAME, None)
+        self.assertFalse(t, t)
+
+    def test_skin_uninstalled(self):
+        stool = getattr(self.portal.aq_explicit, 'portal_skins')
+        self.assertFalse('ATContentTypes' in stool)
+
+    def test_uninstalledAllTypes(self):
+        # test that all types are uninstalled well
+        ids = ('Document', 'File', 'Folder', 'Image', 'Link',
+               'News Item', 'Topic', 'Event')
+        ttool = getattr(self.portal.aq_explicit, 'portal_types')
+        for i in ids:
+            self.assertFalse(i in ttool, i)
+
+    def test_quickinstallable(self):
+        # Test, if the Product is available in Quickinstaller.  For Plone < 5
+        # it shouldn't (since it's a core dependency), for Plone >= 5 it should
+        # (where plone.app.contenttypes is installed by default instead).
+        # Only Plone 5 is supported here.
+        try:
+            from Products.CMFPlone.utils import get_installer
+        except ImportError:
+            qi = getattr(self.portal, 'portal_quickinstaller')
+            products = [prod['id'] for prod in qi.listInstalledProducts()]
+            self.assertFalse('ATContentTypes' in products)
+            products = [prod['id'] for prod in qi.listInstallableProducts()]
+            self.assertTrue('ATContentTypes' in products)
+        else:
+            qi = get_installer(self.portal)
+            self.assertFalse(
+                qi.is_product_installed('Products.ATContentTypes'))
+            self.assertTrue(
+                qi.is_product_installable('Products.ATContentTypes'))
+
+    def test_removes_related_items_catalog_index(self):
+        self.assertNotIn('getRawRelatedItems', self.cat.Indexes)
+
+tests.append(TestUninstallation)
 
 
 def test_suite():